update dependencies, throttle down paralellism to 3

+++ js/build.js
@@ -7,14 +7,19 @@ console.time('build')
 console.log('BEGIN')
 var [,,source, domain] = process.argv,
 	childProcesses = [],
-	killChilds = () => (childProcesses||[]).forEach(cp => {
-		try {
-			process.kill(-cp.pid)
-		} catch (e){/**/}
-	})
+	killChilds = () => {
+		(childProcesses||[]).forEach(cp => {
+			try {
+				if(cp!=='KILLED'){
+					process.kill(-cp.pid)
+					cp.kill('SIGINT')
+				}
+			} catch (e){/**/}
+		})
+		childProcesses[0] = 'KILLED'
+	}
 
 process.on('error', killChilds)
-process.on('exit', killChilds)
 process.on('SIGINT', killChilds)
 
 loadConfig(source, domain).then(config => {
